review renamed package back to ephemeral_storage
