+++ android/java/org/chromium/chrome/browser/app/BraveActivity.java
@@ -136,8 +136,8 @@ import org.chromium.chrome.browser.crypto_wallet.model.CryptoAccountTypeInfo;
 import org.chromium.chrome.browser.crypto_wallet.util.Utils;
 import org.chromium.chrome.browser.customtabs.CustomTabActivity;
 import org.chromium.chrome.browser.customtabs.FullScreenCustomTabActivity;
-import org.chromium.chrome.browser.first_party_storage_cleaner.BraveFirstPartyStorageCleanerUtils;
-import org.chromium.chrome.browser.first_party_storage_cleaner.FirstPartyStorageCleanerInterface;
+import org.chromium.chrome.browser.ephemeral_storage.BraveFirstPartyStorageCleanerUtils;
+import org.chromium.chrome.browser.ephemeral_storage.FirstPartyStorageCleanerInterface;
 import org.chromium.chrome.browser.flags.ChromeFeatureList;
 import org.chromium.chrome.browser.flags.ChromeSwitches;
 import org.chromium.chrome.browser.fullscreen.BrowserControlsManager;
