+++ browser/ephemeral_storage/android/java/src/org/chromium/chrome/browser/ephemeral_storage/BraveFirstPartyStorageCleanerUtils.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.first_party_storage_cleaner;
+package org.chromium.chrome.browser.ephemeral_storage;
 
 import org.jni_zero.JNINamespace;
 import org.jni_zero.NativeMethods;
